포인트 충전 동시성제어 통합 테스트로 수정 - synchronized 사용 39sec
